requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
cffi==1.15.1
click==8.1.6
colorama==0.4.6
cryptography==41.0.3
dnspython==2.3.0
ecdsa==0.18.0
fastapi==0.100.0
h11==0.14.0
idna==3.4
mysql-connector-python==8.0.33
passlib==1.7.4
#protobuf==4.21.12
protobuf==3.20.3
pyasn1==0.5.0
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
pymongo==4.5.0
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.20.0
